backend/requirements.txt
Django>=5.2
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-storages>=1.14.0
drf-spectacular>=0.27.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.9
weasyprint>=60.0
python-dotenv>=1.0.0
Pillow>=10.0.0
qrcode[pil]>=7.4.0
boto3>=1.34.0
django-filter>=23.5
openpyxl>=3.1.0
xlsxwriter>=3.2.0
Faker>=22.0.0
